Refactor environment and project services; add project initialization tests
- Updated `xcconfig_import_command_test.dart` to use `ProjectService` for project management and removed project name arguments from test cases.
- Created `init_command_test.dart` to test project initialization, including handling of existing project names and initialization without a name.
- Modified `project.dart` and `project_metadata.dart` to include a unique identifier for projects.
- Enhanced `environment_service.dart` to improve error handling and file path management.
- Updated `project_registry_service.dart` and `project_service.dart` to use project IDs for registration and management.
- Added validation for project names in `project_service.dart` to ensure they meet length and character requirements.
- Improved test coverage for environment and project services, including handling of non-existent files and project creation from the current directory.

+++ packages/secure_env_cli/src/cli/commands/env/delete_command.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:mason_logger/mason_logger.dart';
 import 'package:secure_env_core/src/services/environment_service.dart';
 import '../base_command.dart';
@@ -6,18 +8,9 @@ import '../base_command.dart';
 class DeleteCommand extends BaseCommand {
   DeleteCommand({
     required super.logger,
-    EnvironmentService? environmentService,
-  }) : _environmentService = environmentService ??
-            EnvironmentService(
-              logger: logger,
-            ) {
+    required super.projectService,
+  }) {
     argParser
-      ..addOption(
-        'project',
-        abbr: 'p',
-        help: 'Project name',
-        mandatory: true,
-      )
       ..addOption(
         'name',
         abbr: 'n',
@@ -26,7 +19,7 @@ class DeleteCommand extends BaseCommand {
       );
   }
 
-  final EnvironmentService _environmentService;
+  late final EnvironmentService _environmentService;
 
   @override
   String get description => 'Delete an environment';
@@ -36,12 +29,17 @@ class DeleteCommand extends BaseCommand {
 
   @override
   Future<int> run() => handleErrors(() async {
-        final projectName = argResults!['project'] as String;
+        final project = await projectService.getProjectFromCurrentDirectory();
+        if (project == null) {
+          throw 'No project found in the current directory. Please run "secure_env init" first.';
+        }
+        _environmentService = await EnvironmentService.forProject(
+            project: project, projectService: projectService, logger: logger);
+
         final envName = argResults!['name'] as String;
 
         await _environmentService.deleteEnvironment(
           name: envName,
-          projectName: projectName,
         );
 
         logger.success('Deleted environment: $envName');
